# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "企业知识库问答系统 API",
    "version": "1.0.0",
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "本地开发"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "UserId": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "登录用户的 MongoDB _id"
      },
      "UserName": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-name",
        "description": "登录用户名"
      },
      "UserRole": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-role",
        "description": "用户角色：admin 或 user"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer"
          },
          "msg": {
            "type": "string"
          },
          "error": {
            "type": "string"
          }
        }
      },
...
      "AskRequest": {
        "type": "object",
        "required": [
          "question"
        ],
        "properties": {
          "question": {
            "type": "string",
            "example": "如何申请年假？"
          },
          "categoryId": {
            "type": "string",
            "description": "知识库分类 ID（可选）"
          },
          "sessionId": {
            "type": "string",
            "description": "会话 ID，用于续接对话（可选）"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/qa/ask": {
      "post": {
        "tags": [
          "QA"
        ],
        "summary": "流式问答（SSE）",
        "description": "核心问答接口，返回 Server-Sent Events 流式响应（`text/event-stream`）。\n\n**事件类型：**\n- `sources` — 召回的参考文档列表\n- `mismatch` — 当前知识库与问题不匹配，建议切换\n- `chunk` — LLM 生成的文本片段\n- `done` — 生成完成，包含完整答案和 sessionId\n- `error` — 发生错误\n\n详细协议说明见文档首页「SSE 流式问答协议」章节。\n",
        "security": [
          {
            "UserId": [],
            "UserName": [],
            "UserRole": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "SSE 流式响应",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "event: sources\ndata: [{\"documentId\":\"...\",\"title\":\"员工手册\"}]\n\nevent: chunk\ndata: \"根据\"\n\nevent: done\ndata: {\"answer\":\"完整答案\",\"sessionId\":\"...\"}\n"
                }
              }
            }
          },
          "400": {
            "description": "问题不能为空",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "问答失败"
          }
        }
      }
    },
...
  }
}